Move deployments to separate app
