+++ api/resources.py
@@ -16,9 +16,6 @@ from .models import (
     Profile,
     ActionsTaken,
     Action,
-    ERUOwner,
-    ERU,
-    Heop,
 )
 from .authentication import ExpiringApiKeyAuthentication
 from .authorization import FieldReportAuthorization, UserProfileAuthorization
@@ -296,55 +293,3 @@ class FieldReportResource(ModelResource):
             'gov_num_displaced',
             'gov_num_assisted',
         ]
-
-
-
-class ERUOwnerResource(ModelResource):
-    eru_set = fields.ToManyField('api.resources.ERUResource', 'eru_set', null=True, full=True)
-    class Meta:
-        queryset = ERUOwner.objects.all()
-        authentication = ExpiringApiKeyAuthentication()
-        resource_name = 'eru_owner'
-        allowed_methods = ['get']
-        filtering = {
-            'country': ('exact', 'in'),
-        }
-
-
-class RelatedERUOwnerResource(ModelResource):
-    country = fields.ForeignKey(CountryResource, 'country', full=True)
-    class Meta:
-        queryset = ERUOwner.objects.all()
-        allowed_methods = ['get']
-
-
-class ERUResource(ModelResource):
-    countries = fields.ToManyField(CountryResource, 'countries', full=True, null=True)
-    eru_owner = fields.ForeignKey(RelatedERUOwnerResource, 'eru_owner', full=True)
-    class Meta:
-        queryset = ERU.objects.all()
-        authentication = ExpiringApiKeyAuthentication()
-        filtering = {
-            'eru_owner': ALL_WITH_RELATIONS,
-            'type': ('exact', 'in'),
-            'countries': ('in'),
-        }
-        allowed_methods = ['get']
-
-
-class HeopResource(ModelResource):
-    country = fields.ForeignKey(CountryResource, 'country', full=True)
-    region = fields.ForeignKey(RegionResource, 'region', full=True)
-    event = fields.ForeignKey(EventResource, 'event', null=True)
-    dtype = fields.ForeignKey(DisasterTypeResource, 'dtype', null=True, full=True)
-    class Meta:
-        queryset = Heop.objects.all()
-        allowed_methods = ['get']
-        authentication = ExpiringApiKeyAuthentication()
-        filtering = {
-            'country': ('exact', 'in'),
-            'region': ('exact', 'in'),
-            'linked_event': ('exact', 'in'),
-            'person': ('exact', 'in'),
-        }
-        ordering = ['end_date', 'start_date']
